Dan | Set up SignUpApp
