--- loginApp/index.html
@@ -1,10 +1,0 @@
-<html>
-<head>
-	<meta charset="utf-8">
-</head>
-
-<body>
-  <div id='app'></div>
-	<script type="text/javascript" src="./bundle.js" charset="utf-8"></script>
-</body>
-</html>
