+++ signupApp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "loginapp",
+  "name": "signupapp",
   "version": "1.0.0",
   "description": "",
   "scripts": {
@@ -17,25 +17,11 @@
   "devDependencies": {
     "babel-core": "^6.4.5",
     "babel-loader": "^6.2.1",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-0": "^6.5.0",
-    "chai": "^3.5.0",
-    "css-loader": "^0.23.1",
     "enzyme": "^2.2.0",
     "jsx-loader": "^0.13.2",
     "mocha": "^3.2.0",
     "sinon": "^1.17.7",
-    "sinon-chai": "^2.8.0",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1"
-  },
-  "babel": {
-    "presets": [
-      "es2015",
-      "stage-0",
-      "react"
-    ]
+    "webpack": "^2.3.0",
+    "webpack-dev-server": "^2.4.2"
   }
 }
