+++ signupApp/webpack.config.js
